+++ src/LionWeb.Core/Core/NodeBase.cs
@@ -111,7 +111,7 @@ public abstract partial class NodeBase : ReadableNodeBase<INode>, INode
         if (value is not IEnumerable)
             throw new InvalidValueException(feature, value);
         var safeNodes = M2Extensions.AsNodes<INode>(value, feature).ToList();
-        AssureAnnotations(safeNodes);
+        AssureAnnotations(M2Extensions.AsNodes<IAnnotationInstance>(safeNodes, feature).ToList());
         AnnotationSetNotificationEmitter notification = new(this, safeNodes, _annotations);
         notification.CollectOldData();
         RemoveSelfParent(_annotations.ToList(), _annotations, null);
@@ -126,7 +126,7 @@ public abstract partial class NodeBase : ReadableNodeBase<INode>, INode
     /// <inheritdoc />
     public virtual void AddAnnotations(IEnumerable<INode> annotations)
     {
-        var safeAnnotations = AssureAnnotations(annotations?.ToList());
+        var safeAnnotations = AssureAnnotations(M2Extensions.AsNodes<IAnnotationInstance>(annotations, null).ToList());
         int index = Math.Max(_annotations.Count - 1, 0);
         foreach (var safeAnnotation in safeAnnotations)
         {
@@ -157,7 +157,7 @@ public abstract partial class NodeBase : ReadableNodeBase<INode>, INode
     public virtual void InsertAnnotations(Index index, IEnumerable<INode> annotations)
     {
         AssureInRange(index, _annotations);
-        var safeAnnotations = AssureAnnotations(annotations?.ToList());
+        var safeAnnotations = AssureAnnotations(M2Extensions.AsNodes<IAnnotationInstance>(annotations, null).ToList());
         foreach (var safeAnnotation in safeAnnotations)
         {
             AnnotationAddMultipleNotificationEmitter notification = new(this, safeAnnotation, _annotations,