+++ src/LionWeb.Generator/GeneratorExtensions/Extensions.cs
@@ -17,7 +17,10 @@
 
 namespace LionWeb.Generator.GeneratorExtensions;
 
+using Core.M2;
+using Core.M3;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
@@ -54,5 +57,134 @@ public static class Extensions
                 onPersistedFile?.Invoke(copy, file);
             }
         }
+    }
+
+    /// <inheritdoc cref="AddUsing(CompilationUnitSyntax, string)"/>
+    public static CompilationUnitSyntax AddUsing(this CompilationUnitSyntax compilationUnit, Type type) =>
+        compilationUnit.AddUsing(type.Namespace!);
+
+    /// <summary>
+    /// Adds "global::<i><paramref name="nameSpace"/></i>" to the <i>usings</i> of the first namespace in <paramref name="compilationUnit"/>.
+    /// </summary>
+    /// <returns>Modified version of <paramref name="compilationUnit"/>.</returns>
+    public static CompilationUnitSyntax AddUsing(this CompilationUnitSyntax compilationUnit, string nameSpace)
+    {
+        var ns = compilationUnit.Members.OfType<FileScopedNamespaceDeclarationSyntax>().First();
+        return compilationUnit.ReplaceNode(
+            ns,
+            ns.AddUsings(UsingDirective(ParseName($"global::{nameSpace}")))
+        );
+    }
+
+    /// <summary>
+    /// Adds <paramref name="annotations"/> to the initializers in the language constructor in <paramref name="compilationUnit"/>.
+    /// </summary>
+    /// <returns>Modified version of <paramref name="compilationUnit"/>.</returns>
+    public static CompilationUnitSyntax AddM2Annotations(this CompilationUnitSyntax compilationUnit,
+        Correlator correlator, Dictionary<IKeyed, CollectionExpressionSyntax> annotations)
+    {
+        var correlations = annotations.Keys
+            .SelectMany(correlator.FindAll<KeyedToLanguageConstructorCorrelation>);
+
+        foreach (var correlation in correlations)
+        {
+            var assignment = correlation.LookupIn(compilationUnit);
+            if (assignment.Right is ImplicitObjectCreationExpressionSyntax
+                {
+                    ArgumentList.Arguments:
+                    [
+                        {
+                            Expression: ParenthesizedLambdaExpressionSyntax
+                            {
+                                Body: ObjectCreationExpressionSyntax
+                                {
+                                    Initializer: { } init
+                                }
+                            }
+                        }
+                    ]
+                })
+            {
+                compilationUnit = compilationUnit.ReplaceNode(
+                    init,
+                    init.AddExpressions(AssignmentExpression(
+                        SyntaxKind.SimpleAssignmentExpression,
+                        IdentifierName("Annotations"),
+                        annotations[correlation.Keyed]
+                    ))
+                );
+            }
+        }
+
+        return compilationUnit;
+    }
+
+    /// <summary>
+    /// Adds `sealed` keyword to all classes inside <paramref name="compilationUnit"/>
+    /// that are generated from non-specialized <see cref="Classifier"/>s.
+    /// Looks through all <see cref="languages"/> to find specializations. 
+    /// </summary>
+    /// <returns>Modified version of <paramref name="compilationUnit"/>.</returns>
+    public static CompilationUnitSyntax SealClassifiers(this CompilationUnitSyntax compilationUnit,
+        Correlator correlator, List<Language> languages, ILookup<Classifier, Classifier>? directSpecializations = null)
+    {
+        directSpecializations ??= M2Extensions.MapDirectSpecializations(languages);
+
+        var classifiersWithoutSpecializations = languages
+            .SelectMany(l => l.Entities)
+            .OfType<Classifier>()
+            .Where(c => c switch
+            {
+                Concept { Abstract: true } => false,
+                Interface => false,
+                _ => true
+            })
+            .Where(c => !c.AllSpecializations(languages, directSpecializations: directSpecializations).Any())
+            .SelectMany(correlator.FindAll<IClassifierToMainCorrelation>);
+
+        foreach (var correlation in classifiersWithoutSpecializations)
+        {
+            var typeDeclaration = correlation.LookupIn(compilationUnit);
+            compilationUnit = compilationUnit.ReplaceNode(
+                typeDeclaration,
+                typeDeclaration
+                    .WithModifiers(TokenList(
+                        typeDeclaration.Modifiers.Append(Token(SyntaxKind.SealedKeyword))
+                            .Order(new TokenComparer())
+                    ))
+            );
+        }
+
+        return compilationUnit;
+    }
+
+    private class TokenComparer : IComparer<SyntaxToken>
+    {
+        private static readonly List<SyntaxKind> _order =
+        [
+            SyntaxKind.PublicKeyword,
+            SyntaxKind.PrivateKeyword,
+            SyntaxKind.ProtectedKeyword,
+            SyntaxKind.InternalKeyword,
+            SyntaxKind.FileKeyword,
+            SyntaxKind.StaticKeyword,
+            SyntaxKind.ExternKeyword,
+            SyntaxKind.NewKeyword,
+            SyntaxKind.VirtualKeyword,
+            SyntaxKind.AbstractKeyword,
+            SyntaxKind.SealedKeyword,
+            SyntaxKind.OverrideKeyword,
+            SyntaxKind.ReadOnlyKeyword,
+            SyntaxKind.UnsafeKeyword,
+            SyntaxKind.RequiredKeyword,
+            SyntaxKind.VolatileKeyword,
+            SyntaxKind.AsyncKeyword,
+            SyntaxKind.PartialKeyword
+        ];
+
+        public int Compare(SyntaxToken x, SyntaxToken y) => Priority(x).CompareTo(Priority(y));
+
+        private static int Priority(SyntaxToken token) =>
+            _order.IndexOf(token.Kind());
     }
 }