+++ test/LionWeb.Generator.Test/PostprocessGeneratorTests.cs
@@ -17,15 +17,18 @@
 
 namespace LionWeb.Generator.Test;
 
-using Core.M2;
+using Core;
+using Core.M1;
 using Core.M3;
-using Core.Test.Languages.Generated.V2023_1.MultiInheritLang;
-using Core.Test.Languages.Generated.V2023_1.TestLanguage;
+using Core.Test.Languages.Generated.V2024_1.MultiInheritLang;
+using Core.Test.Languages.Generated.V2024_1.TestLanguage;
 using GeneratorExtensions;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Formatting;
 using Names;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using Annotation = Core.M3.Annotation;
 
 [TestClass]
 public class PostprocessGeneratorTests
@@ -38,39 +41,88 @@ public class PostprocessGeneratorTests
         var generator = new GeneratorFacade
         {
             Names = new Names(language, "TestLanguage"),
-            LionWebVersion = TestLanguageLanguage.Instance.LionWebVersion
+            LionWebVersion = language.LionWebVersion
+        };
+
+        var compilationUnit = generator.Generate();
+
+        compilationUnit = compilationUnit.SealClassifiers(generator.Correlator, [language]);
+
+        var workspace = new AdhocWorkspace();
+        var generatedContent = Formatter.Format(compilationUnit, workspace).GetText().ToString();
+        Assert.Contains("public sealed partial class CombinedConcept", generatedContent);
+    }
+
+    [TestMethod]
+    public void SealedNoClassifiers()
+    {
+        var lionWebVersion = LionWebVersions.Current;
+        var language = new DynamicLanguage("l", lionWebVersion) { Key = "lang", Name = "lang", Version = "lang" };
+        language.Enumeration("enm", "enm", "enm");
+        language.StructuredDataType("sdt", "sdt", "sdt");
+
+        var generator = new GeneratorFacade
+        {
+            Names = new Names(language, "TestLanguage"),
+            LionWebVersion = language.LionWebVersion
+        };
+
+        var compilationUnit = generator.Generate();
+
+        compilationUnit = compilationUnit.SealClassifiers(generator.Correlator, [language]);
+
+        var generatedContent = compilationUnit.GetText().ToString();
+        Assert.DoesNotContain("sealed", generatedContent);
+    }
+
+    [TestMethod]
+    public void SealedNoMatchingClassifiers()
+    {
+        var lionWebVersion = LionWebVersions.Current;
+        var lang0 = new DynamicLanguage("lang0", lionWebVersion) { Key = "lang0", Name = "lang0", Version = "lang0" };
+        lang0.Interface("iface", "iface", "iface");
+        lang0.Concept("conc", "conc", "conc").IsAbstract(true);
+
+        var lang1 = new DynamicLanguage("lang1", lionWebVersion) { Key = "lang1", Name = "lang1", Version = "lang1" };
+        lang1.Concept("conc", "conc", "conc");
+
+        var generator = new GeneratorFacade
+        {
+            Names = new Names(lang0, "TestLanguage"),
+            LionWebVersion = lang0.LionWebVersion
+        };
+
+        var compilationUnit = generator.Generate();
+
+        compilationUnit = compilationUnit.SealClassifiers(generator.Correlator, [lang0, lang1]);
+
+        var generatedContent = compilationUnit.GetText().ToString();
+        Assert.DoesNotContain("sealed", generatedContent);
+    }
+
+    [TestMethod]
+    public void AddAnnotation()
+    {
+        var language = TestLanguageLanguage.Instance;
+
+        var generator = new GeneratorFacade
+        {
+            Names = new Names(language, "TestLanguage"), LionWebVersion = language.LionWebVersion
         };
 
         var compilationUnit = generator.Generate();
         var correlator = generator.Correlator;
 
-        var classifiersWithoutSpecializations = language.Entities
-            .OfType<Classifier>()
-            .Where(c => !c.AllSpecializations([language]).Any())
-            .Select(c => correlator.FindAll<IClassifierToMainCorrelation>(c).Single());
+        int i = 0;
+        var annotations = M1Extensions.Descendants<IKeyed>(language)
+            .OfType<Annotation>()
+            .ToDictionary(IKeyed (k) => k,
+                v => (CollectionExpressionSyntax)ParseExpression($"""[new TestAnnotation("testAnnotation{i++}")]""")
+                );
 
-        foreach (var correlation in classifiersWithoutSpecializations)
-        {
-            var typeDeclaration = correlation.LookupIn(compilationUnit);
-            compilationUnit = compilationUnit.ReplaceNode(
-                typeDeclaration,
-                typeDeclaration
-                    .AddModifiers(Token(SyntaxKind.SealedKeyword))
-            );
-        }
+        compilationUnit = compilationUnit.AddM2Annotations(correlator, annotations);
 
-        var path = Path.GetTempFileName();
-        try
-        {
-            generator.Persist(path, compilationUnit);
-            Console.WriteLine($"Wrote output to {path}");
-
-            var generatedContent = File.ReadAllText(path);
-            Assert.Contains("public partial sealed class CombinedConcept", generatedContent);
-        } finally
-        {
-            if (Path.Exists(path))
-                File.Delete(path);
-        }
+        var generatedContent = compilationUnit.GetText().ToString();
+        Assert.Contains("""[new TestAnnotation("testAnnotation0")]""", generatedContent);
     }
 }