+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
 
 ## [0.4.0] - tbd
 ### Added
+* Introduced `NodeBuilder` to emit Roslyn AST that creates a LionWeb node instance. 
+* Added Generator extension to mark non-specialized Classifiers as `sealed`.
+* Added Generator extension to add annotations to generated M2 types.
+* Enable attaching annotations to generated M2 types.
+* Added `M1Extensions.WithAnnotation()` to fluently add an annotation to a node.
 * Added `GeneratorFacade` extension `PersistFilePerType()` to generate one file per generated type.
 * `INewNodeNotification` and `IDeletedNodeNotification` to mark notifications that add/remove nodes from the (notified) world.
 * `GeneratorFacade` now exposes `Correlations` property. This correlates LionWeb language elements to their corresponding C# syntax nodes.
@@ -25,6 +30,7 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
   The multiple mutators (`AddAnnotationsRaw()`, `InsertContainmentsRaw()`, `RemoveReferencesRaw()`, etc.) take only element instead of a list as most use cases work with single elements.
 * Added test to assure changed base classes are compatible with previously generated language
 ### Fixed
+* List Primitive datatypes as entities of generated languages. 
 * Bug fix: Introduce check to validate that the deleted node's ID matches the actual node's ID during child deletion notification.
 * Bug fix: Notification emitter now checks whether a moved node is part of partition of not. If the moved node is a floating node – i.e. not part of a partition – and it replaces another node, then `ChildReplacedNotification` is emitted. If there is no replacement, then `ChildAddedNotification` is emitted.
 * Bugfix: `DeltaDeserializerHandler` is made public. This handler enables deserializer to accept node id that appears both in received delta(s) and local nodes. In the context of delta protocol, this enables replacing a node in a model with a new node with the same id, which results in a valid model.